Commit rename de componentes
