+++ src/App.tsx
@@ -1,16 +1,16 @@
-import Button_Class from "@components/Button_Class";
-import Progress_Chart from "@components/Progress_Chart";
+import ButtonClass from "@components/ButtonClass";
+import ProgressChart from "@components/ProgressChart";
 function App() {
   return (
     <div className="App">
-      <Button_Class
+      <ButtonClass
         title="Diseño Web"
         description="Diseño web es una disciplina que se enfoca en la creación y diseño de sitios web, abarcando aspectos como la estructura, la apariencia visual, la usabilidad y la interacción."
         homework="Avance EC1"
         teacherName="JESUS ALBERTO OJEDA SAUCEDO"
       />
 
-      <Progress_Chart percentage="70"></Progress_Chart>
+      <ProgressChart percentage="70"></ProgressChart>
     </div>
   );
 }
