+++ src/theme/styles.scss
@@ -287,7 +287,7 @@ img#iconGoogle {
 // ################################################################################################ PROGRESS CHART
 // No esta con SCSS, falta optimización
 $--clr: #2f8da1;
-$--percentaje: 5;
+$--percentaje: 88;
 .container-chart {
   position: relative;
   display: flex;
